requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib==1.7.4
bcrypt==3.2.2
